requirements.txt
bcrypt 
mysql-connector-python.